requirements.txt
pytorch==2.0.0
numpy==1.23.5
tqdm==4.66.1
scikit-learn==1.3.2
pillow==10.0.1